requirements.txt
numpy~=1.26.2
Pillow~=10.1.0

matplotlib~=3.8.2
seaborn
pycocotools~=2.0.7

scipy~=1.11.4
tensorflow~=2.15.0

keras~=2.15.0
